refactored header menu into separate file updated useRef vars and passed to menu components
